:recycle: Screen files move to route folder
