+++ order/composeApp/src/commonMain/kotlin/io/reyurnible/order/ui/routes/order_complete/OrderCompleteRoute.kt
@@ -2,7 +2,6 @@ package io.reyurnible.order.ui.routes.order_complete
 
 import androidx.compose.runtime.Composable
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import io.reyurnible.order.ui.screens.OrderCompleteScreen
 
 @Composable
 fun OrderCompleteRoute(
