+++ order/composeApp/src/commonMain/kotlin/io/reyurnible/order/ui/routes/order_complete/OrderCompleteUiState.kt
@@ -1,0 +1,7 @@
+package io.reyurnible.order.ui.routes.order_complete
+
+// Visible screen state
+data class OrderCompleteUiState(
+    val isLoading: Boolean = false,
+    val error: Throwable? = null
+)
