+++ order/composeApp/src/commonMain/kotlin/io/reyurnible/order/ui/routes/order_confirmation/OrderConfirmationRoute.kt
@@ -2,7 +2,6 @@ package io.reyurnible.order.ui.routes.order_confirmation
 
 import androidx.compose.runtime.Composable
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import io.reyurnible.order.ui.screens.OrderConfirmationScreen
 
 @Composable
 fun OrderConfirmationRoute(
