+++ order/composeApp/src/commonMain/kotlin/io/reyurnible/order/ui/routes/order_confirmation/OrderConfirmationUiState.kt
@@ -1,0 +1,7 @@
+package io.reyurnible.order.ui.routes.order_confirmation
+
+// Visible screen state
+data class OrderConfirmationUiState(
+    val isLoading: Boolean = false,
+    val error: Throwable? = null
+)
