+++ order/composeApp/src/commonMain/kotlin/io/reyurnible/order/ui/routes/order_history/OrderHistoryUiState.kt
@@ -1,0 +1,7 @@
+package io.reyurnible.order.ui.routes.order_history
+
+// Visible screen state
+data class OrderHistoryUiState(
+    val isLoading: Boolean = false,
+    val error: Throwable? = null
+)
