+++ order/composeApp/src/commonMain/kotlin/io/reyurnible/order/ui/routes/order_history/OrderHistoryViewModel.kt
@@ -8,12 +8,6 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
 
-// Visible screen state
-data class OrderHistoryUiState(
-    val isLoading: Boolean = false,
-    val error: Throwable? = null
-)
-
 // Inner state
 private data class OrderHistoryViewModelState(
     val isLoading: Boolean = false,
