+++ order/composeApp/src/commonMain/kotlin/io/reyurnible/order/ui/routes/select_item/SelectItemRoute.kt
@@ -2,9 +2,6 @@ package io.reyurnible.order.ui.routes.select_item
 
 import androidx.compose.runtime.Composable
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import io.reyurnible.order.ui.screens.OrderCompleteScreen
-import io.reyurnible.order.ui.screens.OrderConfirmationScreen
-import io.reyurnible.order.ui.screens.SelectItemScreen
 
 @Composable
 fun SelectItemRoute(
