+++ order/composeApp/src/commonMain/kotlin/io/reyurnible/order/ui/routes/select_item/SelectItemScreen.kt
@@ -1,4 +1,4 @@
-package io.reyurnible.order.ui.screens
+package io.reyurnible.order.ui.routes.select_item
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
@@ -13,6 +13,7 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Button
 import androidx.compose.material.Card
@@ -23,24 +24,27 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import io.reyurnible.order.ui.components.OrderAppBar
-import io.reyurnible.order.ui.routes.order_history.OrderHistoryUiState
 import order.composeapp.generated.resources.Res
 import order.composeapp.generated.resources.compose_multiplatform
-import order.composeapp.generated.resources.order_history__title
 import order.composeapp.generated.resources.select_item__title
 import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.ui.tooling.preview.Preview
+import kotlin.jvm.JvmInline
 
 @Composable
-fun OrderHistoryScreen(
-    uiState: OrderHistoryUiState,
-    onCheckoutButtonClicked: () -> Unit
+fun SelectItemScreen(
+    uiState: SelectItemUiState,
+    onOrderConfirmButtonClicked: () -> Unit,
+    onItemClickPlusItem: (ItemId) -> Unit = {},
+    onItemClickMinusItem: (ItemId) -> Unit = {},
+    onItemClickAddToCart: (ItemId) -> Unit = {},
 ) {
     Scaffold(
         topBar = {
             OrderAppBar(
-                currentScreenName = Res.string.order_history__title,
+                currentScreenName = Res.string.select_item__title,
                 canNavigateBack = false,
+
             )
         }
     ) { innerPadding ->
@@ -50,14 +54,13 @@ fun OrderHistoryScreen(
                 contentPadding = PaddingValues(horizontal = 16.dp, vertical = 16.dp),
                 verticalArrangement = Arrangement.spacedBy(8.dp),
             ) {
-                item {
-                    Text("注文履歴")
-                    Spacer(modifier = Modifier.padding(8.dp))
-                    Text("合計金額: 10000円")
-                    Spacer(modifier = Modifier.padding(8.dp))
-                }
-                items(10) { index ->
-                    OrderDetailRow(index)
+                items(uiState.selectItem) { item ->
+                    ItemRow(
+                        item = item,
+                        onClickPlusItem = { onItemClickPlusItem(it) },
+                        onClickMinusItem = { onItemClickMinusItem(it) },
+                        onClickAddToCart = { onItemClickAddToCart(it) },
+                    )
                 }
             }
             Box(
@@ -66,18 +69,32 @@ fun OrderHistoryScreen(
                 Button(
                     modifier = Modifier.fillMaxWidth(),
                     shape = RoundedCornerShape(24.dp),
-                    onClick = { onCheckoutButtonClicked() },
+                    onClick = { onOrderConfirmButtonClicked() },
                 ) {
-                    Text("お会計する")
+                    Text("注文内容を確認する")
                 }
             }
         }
     }
 }
 
+@JvmInline
+value class ItemId(val id: Int)
+
+data class SelectItem(
+    val id: ItemId,
+    val name: String,
+    val price: Int,
+    val imageUrl: String,
+    val currentItemCount: Int,
+)
+
 @Composable
-fun OrderHistoryRow(
-    index: Int
+fun ItemRow(
+    item: SelectItem,
+    onClickPlusItem: (ItemId) -> Unit = {},
+    onClickMinusItem: (ItemId) -> Unit = {},
+    onClickAddToCart: (ItemId) -> Unit = {},
 ) {
     Card(
         modifier = Modifier
@@ -91,10 +108,22 @@ fun OrderHistoryRow(
                 modifier = Modifier.size(64.dp, 64.dp),
             )
             Column(modifier = Modifier.wrapContentSize().padding(8.dp)) {
-                Text("商品名: Item $index")
+                Text("商品名: ${item.name}")
                 Spacer(modifier = Modifier.padding(8.dp))
-                Text("価格: ${index * 1000}円")
-                Text("個数: 1個")
+                Text("価格: ${item.price}円")
+                Row(verticalAlignment = Alignment.CenterVertically) {
+                    Button(onClick = { onClickMinusItem.invoke(item.id) }) {
+                        Text("-")
+                    }
+                    Text("${item.currentItemCount}", Modifier.padding(8.dp))
+                    Button(onClick = { onClickPlusItem.invoke(item.id) }) {
+                        Text("+")
+                    }
+                    Spacer(modifier = Modifier.fillMaxWidth().weight(1f))
+                    Button(onClick = { onClickAddToCart.invoke(item.id) }) {
+                        Text("カートに追加")
+                    }
+                }
             }
         }
     }
@@ -102,6 +131,14 @@ fun OrderHistoryRow(
 
 @Composable
 @Preview
-fun OrderHistoryRowPreview() {
-    OrderHistoryRow(1)
+fun ItemRowPreview() {
+    ItemRow(
+        SelectItem(
+        id = ItemId(1),
+        name = "商品名",
+        price = 1000,
+        imageUrl = "",
+        currentItemCount = 1
+    )
+    )
 }
