+++ order/composeApp/src/commonMain/kotlin/io/reyurnible/order/ui/routes/select_item/SelectItemUiState.kt
@@ -1,0 +1,8 @@
+package io.reyurnible.order.ui.routes.select_item
+
+// Visible screen state
+data class SelectItemUiState(
+    val isLoading: Boolean = false,
+    val error: Throwable? = null,
+    val selectItem: List<SelectItem>,
+)
