+++ order/composeApp/src/commonMain/kotlin/io/reyurnible/order/ui/routes/start/StartRoute.kt
@@ -2,10 +2,6 @@ package io.reyurnible.order.ui.routes.start
 
 import androidx.compose.runtime.Composable
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import io.reyurnible.order.ui.screens.OrderCompleteScreen
-import io.reyurnible.order.ui.screens.OrderConfirmationScreen
-import io.reyurnible.order.ui.screens.SelectItemScreen
-import io.reyurnible.order.ui.screens.StartScreen
 
 @Composable
 fun StartRoute(
@@ -14,13 +10,10 @@ fun StartRoute(
 ) {
     val uiState = viewModel.uiState.collectAsStateWithLifecycle()
 
-    when(uiState.value) {
-        is StartUiState.Start -> {
-            StartScreen(
-                onStartButtonClicked = {
-                    viewModel.onStartButtonClicked(onSubmitSuccess = onSubmitSuccess)
-                }
-            )
+    StartScreen(
+        uiState = uiState.value,
+        onStartButtonClicked = {
+            viewModel.onStartButtonClicked(onSubmitSuccess = onSubmitSuccess)
         }
-    }
+    )
 }
