+++ order/composeApp/src/commonMain/kotlin/io/reyurnible/order/ui/routes/start/StartScreen.kt
@@ -1,4 +1,4 @@
-package io.reyurnible.order.ui.screens
+package io.reyurnible.order.ui.routes.start
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -7,7 +7,6 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.widthIn
-import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Button
 import androidx.compose.material.Scaffold
@@ -16,25 +15,13 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import io.reyurnible.order.ui.components.OrderAppBar
-import io.reyurnible.order.ui.routes.order_complete.OrderCompleteUiState
-import order.composeapp.generated.resources.Res
-import order.composeapp.generated.resources.order_complete__title
-import order.composeapp.generated.resources.order_history__title
 
 @Composable
-fun OrderCompleteScreen(
-    uiState: OrderCompleteUiState,
-    onBackButtonClicked: () -> Unit,
+fun StartScreen(
+    uiState: StartUiState,
+    onStartButtonClicked: () -> Unit,
 ) {
-    Scaffold(
-        topBar = {
-            OrderAppBar(
-                currentScreenName = Res.string.order_complete__title,
-                canNavigateBack = false,
-            )
-        }
-    ) {
+    Scaffold {
         Box(
             modifier = Modifier
                 .fillMaxSize()
@@ -44,14 +31,14 @@ fun OrderCompleteScreen(
             Column(
                 horizontalAlignment = Alignment.CenterHorizontally
             ) {
-                Text("ご注文ありがとうございます！")
+                Text("ご来店ありがとうございます！")
                 Spacer(modifier = Modifier.padding(24.dp))
                 Button(
                     modifier = Modifier.widthIn(max = 280.dp).fillMaxWidth(),
                     shape = RoundedCornerShape(24.dp),
-                    onClick = { onBackButtonClicked() },
+                    onClick = { onStartButtonClicked() },
                 ) {
-                    Text("注文画面に戻る")
+                    Text("Start")
                 }
             }
         }
