+++ order/composeApp/src/commonMain/kotlin/io/reyurnible/order/ui/routes/start/StartUiState.kt
@@ -1,0 +1,6 @@
+package io.reyurnible.order.ui.routes.start
+
+class StartUiState(
+    val isLoading: Boolean = false,
+    val error: Throwable? = null
+)
